+++ src/px-script/dialog/index.js
@@ -19,6 +19,10 @@ class Dialog {
             });
         }
 
+        if (this.isOpen) {
+            document.body.classList.add("dialog-open");
+        }
+
         // Close dialog on esc
         document.addEventListener("keydown", e => {
             e.keyCode === 27 ? this.close() : null;
@@ -31,60 +35,60 @@ class Dialog {
     }
 
     open () {
-        if (!this.isOpen) {
-            this.isOpen = true;
-            this._el.classList.add("d-flex");
-            document.body.classList.add("dialog-open");
-        }
+        this.isOpen = true;
+        this._el.classList.add("d-flex");
+        document.body.classList.add("dialog-open");
     }
 
     close () {
-        if (this.isOpen) {
-            this.isOpen = false;
-            this._el.classList.remove("d-flex");
-            document.body.classList.remove("dialog-open");
-        }
+        this.isOpen = false;
+        this._el.classList.remove("d-flex");
+        document.body.classList.remove("dialog-open");
     }
 
 }
 
 const dialog = (() => {
     const init = () => {
-        const dialogs = [...document.querySelectorAll(SELECTORS.DIALOG)].map(dialog => new Dialog(dialog));
+        const dialogEls = [...document.querySelectorAll(SELECTORS.DIALOG)];
 
-        // Init open buttons
-        document.querySelectorAll(SELECTORS.OPEN).forEach(btn => {
-            const id = btn.dataset.dialogOpen;
-            let dialog;
+        if (dialogEls) {
+            const dialogs = dialogEls.map(dialog => new Dialog(dialog));
 
-            dialogs.forEach(d => d.id === id ? dialog = d : null);
+            // Init open buttons
+            document.querySelectorAll(SELECTORS.OPEN).forEach(btn => {
+                const id = btn.dataset.dialogOpen;
+                let dialog;
 
-            if (dialog) {
-                btn.addEventListener("click", e => {
-                    e.preventDefault();
-                    dialog.open();
-                });
-            } else {
-                console.warn(`OPEN: No dialog with with id ${id} was found, make sure the attribute data-dialog-open contains the correct id.`);
-            }
-        });
+                dialogs.forEach(d => d.id === id ? dialog = d : null);
 
-        // Init close buttons
-        document.querySelectorAll(SELECTORS.CLOSE).forEach(btn => {
-            const id = btn.dataset.dialogClose;
-            let dialog;
+                if (dialog) {
+                    btn.addEventListener("click", e => {
+                        e.preventDefault();
+                        dialog.open();
+                    });
+                } else {
+                    console.warn(`OPEN: No dialog with with id ${id} was found, make sure the attribute data-dialog-open contains the correct id.`);
+                }
+            });
 
-            dialogs.forEach(d => d.id === id ? dialog = d : null);
+            // Init close buttons
+            document.querySelectorAll(SELECTORS.CLOSE).forEach(btn => {
+                const id = btn.dataset.dialogClose;
+                let dialog;
 
-            if (dialog) {
-                btn.addEventListener("click", e => {
-                    e.preventDefault();
-                    dialog.close();
-                });
-            } else {
-                console.warn(`CLOSE: No dialog with with id ${id} was found, make sure the attribute data-dialog-close contains the correct id.`);
-            }
-        });
+                dialogs.forEach(d => d.id === id ? dialog = d : null);
+
+                if (dialog) {
+                    btn.addEventListener("click", e => {
+                        e.preventDefault();
+                        dialog.close();
+                    });
+                } else {
+                    console.warn(`CLOSE: No dialog with with id ${id} was found, make sure the attribute data-dialog-close contains the correct id.`);
+                }
+            });
+        }
     };
 
     return { init };