+++ src/px-script/utils/index.js
@@ -1,15 +1,3 @@
-export const applyToEachElement = (elements, fn) => {
-    for (let i = 0, l = elements.length; i < l; i++) {
-        fn(elements[i], i);
-    }
-};
-
-export const findAncestor = (el, cls) => {
-    while ((el = el.parentElement) && !el.classList.contains(cls)); // eslint-disable-line curly
-
-    return el;
-};
-
 export const extendObj = (overWrite, obj, ext) => {
     for (const key in ext) {
         if (overWrite) {
@@ -33,8 +21,6 @@ export const isWithinBoundingBox = (x, y, element) => {
 };
 
 export default {
-    applyToEachElement,
-    findAncestor,
     extendObj,
     isWithinBoundingBox
 };