+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## [0.1x.x] - 2018-xx-xx
+## [0.12.0] - 2018-xx-xx
 
 ### Added
+
+- Added class `.button-group`.
 
 ### Changed
 
@@ -14,5 +16,9 @@
 - Fixed issue with action list icon being visible through topbar.
 - Fixed an issue with topbar component where clicking anchor links does not close the open menu.
 - Fixed an issue with datepicker not displaying properly, or at all, when clicking on label.
+- Fixed an issue where alert close button got properties from the new `<a>` styling (blue border bottom).
+- Fixed a minor issue with toast not resetting its type, if called without a type after another toast with a type has already been called.
 
 ### Removed
+
+- Removed modal from designguide.